garantindo q só o dono da review possa remover ela

+++ src/modules/review/services/DeleteReviewService.ts
@@ -1,6 +1,7 @@
 import { injectable, inject } from "tsyringe";
 import { IReviewsRepository } from "../domain/repositories/IReviewsRepository";
 import { IReview } from "@shared/interfaces/IReview";
+import AppError from "@shared/errors/AppError";
 
 @injectable()
 export default class DeleteReviewService {
@@ -8,7 +9,11 @@ export default class DeleteReviewService {
 		@inject("ReviewsRepository")
 		public reviewsRepository: IReviewsRepository,
 	) {}
-	public execute(id: string): Promise<IReview | null> {
+	public async execute(id: string, pacientId: string): Promise<IReview | null> {
+		const review = await this.reviewsRepository.findById(id);
+		if (review?.pacientId !== pacientId) {
+			throw new AppError("Você não pode apagar um comentário de outra pessoa", 409);
+		}
 		return this.reviewsRepository.remove(id);
 	}
 }
